[4 - 6단계 방탈출 사용자 예약] 카키(서현준) 미션 제출합니다. (#120)
* refactor : 트랜잭션 어노테이션 제거후 명시적으로 데이터 초기화

* refactor : 인기 테마 개수를 파라미터로 받아 조회하도록 변경

* docs : 회원 기능 목록 추가

* feat : 회원 가입 기능 구현

* refactor : 회원가입 url 및 응답 코드 변경

* fix : register 메서드 파라미터 추가 후 오류 수정

* feat : JWT 토큰 생성, 검사 기능 추가

* feat : 회원 로그인 기능 추가

* fix : 전체 테스트 실행 오류 해결

* feat : ArgumentResolver로 로그인 검증 기능 추가

* refactor : 사용자, 관리자 request 명세에 맞게 수정

* feat : 로그인 인터셉터 추가

* fix : 인터셉터에 js 폴더 경로 권한 추가

* refactor : 클라이언트 측 코드에 맞춰 요청 응답 데이터 수정

* refactor : 회원 로그인 기능만 분리해 컨트롤러, 서비스 변경

* feat : 관리자 페이지에 관리자만 접근 가능하도록 권한과, 인터셉터 추가

* test : 프로덕션 코드 변경 사항 테스트에 적용

* feat : 검색 조건 조회 기능 추가

* refactor : NULL 제약 조건 추가

* refactor : 회원 페이지, 어드민 페이지 분리

* refactor : 패키지 구조 변경

* feat : 쿠키를 담당하는 유틸리티 클래스 추가

* refactor : 코드 컨벤션 적용

* refactor : 프로덕션용 dml 추가 및 테스트 설정 분리

* refactor : 미사용 클래스 제거

* test : 회원 이름 테스트 추가

* fix : 회원 아이디에 @NotNull 어노테이션 제거

* refactor : 연결할 DB 드라이버 명 추가

* feat : 예약이 참조하는 테마를 삭제할 경우 예외 발생 기능 추가

* refactor : 예약이 참조하는 시간 조회를 List로 수정 후 참조 여부 확인

* refactor : 정의된 상수 이름으로 값을 비교하도록 수정

* refactor : Login 커스텀 어노테이션 제거

* refactor : 관리자 확인 인터셉터 네이밍 수정

* refactor : 토큰이 없으면 null을 반환하도록 수정

* feat : 토큰이 null일 경우 예외 발생 기능 추가

* feat : 인증 관련 커스텀 예외 추가

* feat : 인증 관련 커스텀 예외 적용 및 검증 테스트 추가

* feat : 401 응답의 에러 페이지 반환 핸들러 추가

* refactor : 일치하는 권한이 없을 경우 미인증 사용자 예외 발생하도록 변경

* test : 쿠키 테스트 추가

* refactor : 불필요한 로직 제거

* refactor : 로그인 회원 조회 url 수정

* refactor : 미인증 사용자는 /members POST 요청만 허용하도록 수정

* fix : member 테이블의 시작 id 수정

* refactor : 예약 조건 검색 DTO를 도메인으로 변경

* refactor : 예약 조건 검색 도메인을 DTO로 변경

* refactor : 컨트롤러 테스트 통합 테스트로 변경

* test : 어드민 권한 확인 테스트 추가

* feat : 로그아웃 기능 추가

* refactor : 코드 컨벤션 적용

+++ src/main/java/roomescape/reservation/service/ThemeService.java
@@ -36,19 +36,24 @@ public class ThemeService {
 
     public List<ThemeResponse> findAll() {
         List<Theme> themes = themeRepository.findAll();
+
         return themes.stream()
                 .map(ThemeResponse::toResponse)
                 .toList();
     }
 
-    public List<PopularThemeResponse> findTopTenThemesDescendingOfLastWeek() {
-        List<Theme> popularTheme = themeRepository.findTopTenThemesDescendingOfLastWeek();
+    public List<PopularThemeResponse> findThemesDescOfLastWeekForLimitCount(int limitCount) {
+        List<Theme> popularTheme = themeRepository.findPopularThemesDescOfLastWeekForLimit(limitCount);
         return popularTheme.stream()
                 .map(PopularThemeResponse::toResponse)
                 .toList();
     }
 
     public void delete(Long id) {
+        List<Theme> themes = themeRepository.findThemesThatReservationReferById(id);
+        if (!themes.isEmpty()) {
+            throw new IllegalArgumentException("해당 테마로 예약된 내역이 있습니다.");
+        }
         themeRepository.delete(id);
     }
 }
